requirements.txt
# Core
langgraph>=0.2.0
langchain-core>=0.3.0
pydantic>=2.0.0

# Image processing
Pillow>=10.0.0

# HTTP client (for API calls)
httpx>=0.27.0
requests>=2.31.0

# Environment
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Long-term memory
# (uses stdlib sqlite3, no extra package needed)

# RAG - vector store + embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Dev / Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=7.0.0
